pyproject.toml
[project]
name = "to-do"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.2",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.13",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.0",
    "pytest>=8.4.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
]
